Updates 1.21.6 support module to 1.21.7
