--- 1.21.6/build.gradle
@@ -1,9 +1,0 @@
-ext {
-    java_version = 21
-}
-
-dependencies {
-    minecraft "com.mojang:minecraft:1.21.6"
-    mappings "net.fabricmc:yarn:1.21.6+build.1:v2"
-    modImplementation "net.fabricmc.fabric-api:fabric-api:0.127.0+1.21.6"
-}
