+++ 1.21.7/build.gradle
@@ -1,0 +1,9 @@
+ext {
+    java_version = 21
+}
+
+dependencies {
+    minecraft "com.mojang:minecraft:1.21.7"
+    mappings "net.fabricmc:yarn:1.21.7+build.1:v2"
+    modImplementation "net.fabricmc.fabric-api:fabric-api:0.128.1+1.21.7"
+}
