+++ 1.21.7/src/main/java/me/videogamesm12/librarian/v1_21_7/addon/FabricAPIAddon.java
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package me.videogamesm12.librarian.v1_21_6.addon;
+package me.videogamesm12.librarian.v1_21_7.addon;
 
 import com.mojang.brigadier.arguments.LongArgumentType;
 import com.mojang.brigadier.arguments.StringArgumentType;
