+++ 1.21.7/src/main/java/me/videogamesm12/librarian/v1_21_7/listeners/ToastNotifier.java
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package me.videogamesm12.librarian.v1_21_6.listeners;
+package me.videogamesm12.librarian.v1_21_7.listeners;
 
 import com.google.common.eventbus.Subscribe;
 import lombok.Getter;
