+++ 1.21.7/src/main/java/me/videogamesm12/librarian/v1_21_7/mixin/CreativeInventoryScreenMixin.java
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package me.videogamesm12.librarian.v1_21_6.mixin;
+package me.videogamesm12.librarian.v1_21_7.mixin;
 
 import com.google.common.eventbus.Subscribe;
 import com.llamalad7.mixinextras.injector.wrapmethod.WrapMethod;
@@ -28,7 +28,7 @@ import me.videogamesm12.librarian.api.event.CacheClearEvent;
 import me.videogamesm12.librarian.api.event.NavigationEvent;
 import me.videogamesm12.librarian.api.event.ReloadPageEvent;
 import me.videogamesm12.librarian.util.ComponentProcessor;
-import me.videogamesm12.librarian.v1_21_6.addon.FabricAPIAddon;
+import me.videogamesm12.librarian.v1_21_7.addon.FabricAPIAddon;
 import net.fabricmc.fabric.impl.client.itemgroup.FabricCreativeGuiComponents;
 import net.kyori.adventure.key.Key;
 import net.kyori.adventure.text.Component;
