+++ 1.21.7/src/main/java/me/videogamesm12/librarian/v1_21_7/mixin/HandledScreenAccessor.java
@@ -15,18 +15,18 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package me.videogamesm12.librarian.v1_21_6.mixin;
+package me.videogamesm12.librarian.v1_21_7.mixin;
 
-import net.minecraft.client.gui.Drawable;
-import net.minecraft.client.gui.screen.Screen;
+import net.minecraft.client.gui.screen.ingame.HandledScreen;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.gen.Accessor;
 
-import java.util.List;
-
-@Mixin(Screen.class)
-public interface ScreenAccessor
+@Mixin(HandledScreen.class)
+public interface HandledScreenAccessor
 {
 	@Accessor
-	List<Drawable> getDrawables();
+	int getX();
+
+	@Accessor
+	int getY();
 }
