+++ 1.21.7/src/main/resources/assets/librarian-resources/icon-256.png
