+++ 1.21.7/src/main/resources/librarian.v1_21_7.mixins.json
@@ -1,7 +1,7 @@
 {
   "required": true,
   "minVersion": "0.8",
-  "package": "me.videogamesm12.librarian.v1_21_6.mixin",
+  "package": "me.videogamesm12.librarian.v1_21_7.mixin",
   "compatibilityLevel": "JAVA_21",
   "mixins": [],
   "injectors": {
