+++ settings.gradle
@@ -41,4 +41,4 @@ include "1.21.1"
 include "1.21.3"
 include "1.21.4"
 include "1.21.5"
-include "1.21.6"
+include "1.21.7"
